+++ tdx-tdcall/Cargo.toml
@@ -20,3 +20,4 @@ x86_64 = { version = "0.14.9", default-features = false, features = ["instructio
 [features]
 default = []
 use_tdx_emulation = []
+tdvmcall = []
